Reorganized structure of auditor.outbound.feign
