+++ auditor.outbound.feign/src/main/java/io/github/tkaczenko/auditor/outbound/feign/api/reader/headers/FeignResponseHeadersHttpReader.java
@@ -1,4 +1,4 @@
-package io.github.tkaczenko.auditor.outbound.feign.reader;
+package io.github.tkaczenko.auditor.outbound.feign.api.reader.headers;
 
 import feign.Response;
 import io.github.tkaczenko.auditor.core.api.reader.headers.HeadersHttpReader;
