+++ demo/src/main/java/io/github/tkaczenko/auditor/demo/config/FeignConfig.java
@@ -1,7 +1,7 @@
 package io.github.tkaczenko.auditor.demo.config;
 
 import feign.Logger;
-import io.github.tkaczenko.auditor.outbound.feign.FeignAuditRequestLogger;
+import io.github.tkaczenko.auditor.outbound.feign.api.FeignAuditRequestLogger;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
